requirements.txt
av==14.4.0
fpdf2==2.8.4
matplotlib==3.10.8
numpy==1.25.2
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
pandas==1.5.3
pandas==2.0.3
Pillow==10.0.1
Pillow==12.1.0
plotly==6.0.1
PyYAML==6.0.2
PyYAML==6.0.3
reportlab==4.4.4
streamlit==1.42.1
streamlit_webrtc==0.64.5
torch==2.6.0+cu124
torch==2.6.0
ultralytics==8.3.114
ultralytics==8.3.102
